Refactor FromBytes and ToBytes for Vec<u8>
Previous version of (de)serializer that for Vec<u8> used generic version for Vec<T>. It was very inefficient because for every u8 element in the vector it would go through the whole process of looking up proper (de)serializer instance, building a Vec<u8> with one element and then coping it to this single element vector. Then, when it was done traversing the input vector it would concatenate all the vectors from Vec<Vec<u8>> into a Vec<u8> doing a lot of wasteful coping and extending in the process.

After injecting gas instructions it was even less efficient and took even more time as every block was followed by two instructions (one instruction that put the cost on the stack and another instruction that made a call to the host charging the gas).

The output of serializing is a Vec<u8> so it is obvious that input Vec<u8> should be serialized as is (prepended with the length of the vector). This changes the serialization for Vec<u8>.

Unfortunately Rust compiler doesn't allow for multiple specializations of the same trait (Vec<T> conflicts with Vec<u8>) so I had to remove Vec<T> implementation and add now missing instance (for Vec<i32> and Vec<Vec<u8>>).