EE-202: Replace Rc<CountingDb> with mutable location in TrackingCopy tests.
In the next commit trackingcopy.rs will be entirely moved out from the storage module and it won't be possible anymore to `impl DbReader for Rc<CountingDb>` because none of the types (`DbReader`, `Rc`) will be defined in the crate (`DbReader` is defined in the storage while the test will be in the engine).

This commit removes the intermediate structure that operates on the `Rc<CountingDb>` and uses mutable location (`Rc<Cell<i32>>`) as a counter for the `CountingDb`.