EE-96: Validate keys in the value to be stored.
When contract wants to `write` value under a key we have to validate
whether it operates with the keys it is allowed to (knows about them).
This is true for both the key it writes to but also about all the keys
that are reachable in the value it wants to store. It means that if the
value to be stored is a `NamedKey` (which is a pair of string and key),
or `Contract` (which holds `known_urefs` map) we have to check whether
these keys are known to the contract that issued the original `write`.