EE-210: Replace `validate_key` implementation with more efficient version.
I had considered various approach before finalizing on the most obvious one.
At first I looked into removing derived `PartialEq`, `Eq` and `Hash`
instances for the `Key` and reimplementing them in a way that ignores the
`AccessRights` of the `URef`. This way I could keep plain `HashSet<Key>`
and check the access rights manually. It would work but it would also
mean that plain equality check of two different `URef`s wouldn't consider
their access rights. This meant that it would be easy to make a mistake
that costs a lot. I decided to not sacrifice safety and UX for the sake
of more elegant encoding of `known_urefs`.