EE-260: Add cache invalidation of read entries in the `TrackingCopyCache`.
Cache invalidation logic is based on the notion of "measurement" and
"maximum cache size". Actual semantics of "measurement" is left for user
to define (by the notion of implementing `Meter` trait). For tests
purposes `Count` is used (which records number of elements put in the cache),
for production code an actual heap size of elements will be used.