EE-260: Replace usage of `lru_cache` create with `linked_hash_map`.
LruCache required to specify a maximum number of elements in a cache.
Since we are invalidating cache based on the "heap size" of its elements
built-in mechanism in the LruCache might have gotten in a way or we had
to specify absurdly high "max_size" when creating LruCache, so that the cache
didn't invalidate elements by itself.

To avoid the problem we use `LinkedHashMap` directly and decide ourselves,
when to remove elements (pop).