EE-345: Add safety checks in mappings.
Method `copy_from_slice` expect source and destination slices to be in
equal size, so ensuring safety at this level is very important as
protobuf doesn't have notion of fixed size bytes type.

Where possible this also changes `clone_from_slice` to `copy_from_slice`
as clone is not necessary for `Copy` types - and we're operating on
`u8` slices only.