EE-345: Fix invalid type error.
Key is defined in `common::key::Key`, but including `ipc::*` also
includes a `Key` which is completely different type. By an explicit
import of `common::key::Key` and removing `ipc::*` (dangerous) wildcard
import and moving to explicit prefixing of IPC types with `ipc::` now
the intention in the code is clearer.