EE-278: Change InsufficientWeight into two errors.
This now has more meaning than just `true`/`false` and leaves room to
actual implementation of logic behind `InsufficientWeight` in future.

This error is meant to still provide consistency with `add_key` and
`remove_key` for contract devs, and in future it can be extended with a
logic to check all deployment key weights.