EE-278: Add `account_dirty` instance for mutating.
This commit adds `account_dirty` to `RuntimeContext` for the purpose of
being able to mutate an account. This instance reflects the state of
`account` as written to the tracking copy instance *before* the
execution begins. This means `account_dirty` has `nonce` updated
already, and this is the same `Account` instance that is kept in the
tracking copy.

Therefore `account_dirty` can be seen as a cache of mutated `Account`
during the execution of a contract.