EE-400: Add helper structure to generate URefs for Genesis setup.
We pregenerate all URefs for the genesis setup process upfront and associate them to their labels (which are Rust constants). This way we can get genesis' URefs by constants. In tests we don't have to simulate generating urefs we don't care about or even track which URef was created when.