EE-468: Added a failing test case for the issue.
Running this test triggers the exploitable deserializing code of vectors
where passed 'length' prefix is not checked for a size, and a
`Vec::with_capacity` is called. First operation on the vector results in
a uncontrolled allocation which ultimately leads to crash of the whole
process:

    memory allocation of 6700148980200 bytes failed
    error: process didn't exit successfully: `./deserialize_error_test-9dd036e11fd2116b --ignored --nocapture` (signal: 6, SIGABRT: process abort signal)

The test case is based on a test case authored by @goral09 originally
introduced in #785.

Co-authored-by: Mateusz Górski <gorski.mateusz@protonmail.ch>