EE-468: Try to reserve space and don't fail.
With this commit, deserializing code for vectors doesn't use
`Vec::with_capacity()` constructor that combined with unsanitized input
which is the length prefix of the serialized data could lead to a crash
of the process.

There is a tradeoff made here that this switches to `try_reserve_exact`
instead of `with_capacity()` - not every size of slice can be compared,
especially if the element of the vector could have dynamic size (such
as Key).

This could be potentially still abused, and attacker could temporaily
allocate hundreds of megs, or even gigs, but it would not lead to a
process crash. For most of the cases where size is correctly serialized
(and not a malicious payload) it will still benefit from the performance
gains.