EE-221: Use unified uref names.
Before, urefs relied on Debug representation of a [u8; 32] type. With
this change, now it is more explicit and uses custom representation of
an URef in the form of `uref-${ADDR}-${ACCESSRIGHTS}`

Where `$ADDR` is base16 representation of the address (64 chars), and
`$ACCESSRIGHTS` is a bit flags encoded in octal, so it occupies max. 3
characters in current representation of it.

This change also shortens the name to just ~73 characters which is also
constant in size.