EE-524: Properly address @henrytill's comment.
This renames the 'placeholder' struct that is used as an error type for
`TryFrom` trait, but also it changes it's signature from `struct Foo` to
`struct Foo(())`. First one can be constructed by anyone through `Foo`,
but the second one has one private field `()` which makes it impossible
to construct it outside of the module.