NO-TICKET: Make ArgParser explicit for single arg.
For quite some time now ArgParser supported a single element to be
treated as a list of arguments with a single element. While it's not
wrong in principle, lack for explicit meaning of that caused whole lot
of troubles that are hard to catch. For example:

```
fn foo(block_hash: [u8; 32], args: impl ArgParser);
```

Both calls would be valid:

```
foo([1; 32], [2; 32]);
```

and

```
foo([2; 32], [1; 32]);
```

And a source of confusion where arguments called contains different
value than expected (i.e. while refactoring, when order of elements
might change).

This change makes the single argument case for `ArgParser` explicit so
you're required to pass single argument as a one element tuple.

```
foo([1;32], [2; 32]); // wrong
foo([1;32], ([2; 32],)); // correct
```