EE-589: Replace PoS finalization test.
Old test tried to execute finalization, but with payment code actually
enabled it turned out to be failing as the payment purse was empty
already, and finalization step was failed just because it was tried to
be executed and observed 0 amount on the rewards purse.

The new test is taking black box approach where it is asserting for
refund purse amount before and after the execution to prove the same
thing.

Co-authored-by: Ed Hastings <ed@casperlabs.io>