NO-TICKET: Don't build tests and doc tests for contracts by default
This PR causes 'cargo test' (and hence 'make test') to not build and run
all the no-op tests and doc tests for the many contracts targets, hence
reducing the build time slightly and decluttering the test output
significantly.

To run these anyway, we can use 'cargo test --all'.

It also adds a cargo config file to the contracts folder in order to
specify the required Wasm build target, which is a more robust approach
than relying on the user or script passing it as a command line arg to
cargo.  This approach is used in integration-testing/contracts.