NODE-923: Make `estimator` accept latest messages as DAG's view.
This is a hotfix to the race condition bug found during simulations (more details in ticket description).
Please note that I had to remove `Casper` trait as signature of `estimator` method did not align in the `MultiParentCasper` (namely the `A` type parameter) was being used as return type of `estimator` and in the `latestMessageHashes` argument (which did not align as `estimator` was returning `List`). I hope this will be fixed in the following PR that addresses the problem correctly.