CON-490: Fix a bug in lmd scoring.
LMD scoring starts with list of latest messages and traverses along the main tree
until the stop hash, collecting scores of each block on the way. If the unfortunate case where
the stop hash was one of the latest messages of a validator AND that stop hash appeared
on the first place in that list, no other message in that list was being followed since
it already reached the stop hash.

The solution is to simply sort that starting list by rank so that stop hash,
which has to be lower in the DAG by the virtue of being LCA, is visited the last.