EE-814: Initial setup for cargo-casperlabs CLI tool
As well as introducing the initial cut of the CLI tool, this commit
removes the Cargo.lock file in preparation for publishing all engine
libraries to crates.io.

The rust-toolchain file was also updated to fix resulting compiler
errors, as were several new lint failures caused by the upgrading the
toolchain.