EE-889: Get rid of serialized public key const.
Renamed to `*_MAX_SIZE` to make it useful for places where we have to
create a vector with `Vec::with_capacity`. The meaning of this constant
is now to contain the maximum possible serialized size.

Where necessary `PUBLIC_KEY_SERIALIZED_LENGTH` is replaced with
additional size parameter coming from the client.