NO-TICKET: Refactor `alloc_bytes` to be stable.
Previously we've used Global trait directly, which proven to be
unstable, and changing often without notice. In this PR I've used
`alloc` function which might be deprecated in future in favor of Global
trait if ever it will become stable. This way we can have stable
function until deprecation notice and further refactor, although it
shouldn't be affected to internal API changes of unstable features.

This also doesn't pass null pointer through FFI, or doesn't try to
allocate zero sized arrays - which according to the docs is UB, but
tries to handle the zero sized allocation at call site by creating empty
vector.

Co-authored-by: Fraser Hutchison <fraser@casperlabs.io>