EE-894: Add support for LTO.
This moves all contract targets to binaries, instead of libraries by
default. This gives the opportunity for LTO to kick in, and reduce size
and execution time of contracts.

This change also clarifies the `feature = "lib"` case, where actual
library is stored in `lib.rs` and the executable in `main.rs` of given
contract.